ui/style: Move the commonlayoutstyle to the style folder
Refs: #2932
