+++ ui/src/components/NodePagePodsTab.js
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux';
 import { useTable } from 'react-table';
 import styled from 'styled-components';
 import { fontSize, padding } from '@scality/core-ui/dist/style/theme';
-import { NodeTab } from './CommonLayoutStyle';
+import { NodeTab } from './style/CommonLayoutStyle';
 import {
   STATUS_RUNNING,
   STATUS_PENDING,
