+++ ui/src/components/style/CommonLayoutStyle.js
